// src/app/core/models/cierre.ts
export class Cierre {
  bizum?: number;
  currentDate?: string;
  efectivo?: number;
  encargada?: string;
  fechaDesde?: string;
  fechaHasta?: string;
  horaDesde?: string;
  horaHasta?: string;
  id?: number;
  idCierre?: number;
  tarjeta?: number;
  total?: number;
  transaccion?: number;
  tratamiento?: number;
}